move confirmEmail to index utils
